Move class Audio to module timeseries
