+++ iracema/__init__.py
@@ -3,8 +3,7 @@ iracema is a python package aimed at the extraction of expressive music
 information from audio signals
 """
 
-from .audio import Audio
-from .timeseries import TimeSeries
+from .timeseries import TimeSeries, Audio
 from .segment import Segment, SegmentList
 
 import iracema.features
